3-Image-Classification-with-CNNs/requirements.txt
numpy==1.21.5
Pillow==9.0.0
torch==1.10.0
torchvision==0.11.1
scikit-learn==1.0.2
matplotlib==3.4.3
pandas==1.3.4
typing-extensions==4.0.1
jupyter==1.0.0
